+++ R/Seurat.R
@@ -5,17 +5,19 @@
 #' @description `to_Seurat()` converts an AnnData object to a Seurat object.
 #' NOTE: Only 1 assay per object is currently. 
 #' @param obj An AnnData object.
+#' @param counts Which layer to use A `layer` within `obj` to use 
 #' @inheritDotParams SeuratObject::CreateSeuratObject
 #'
 #' @importFrom Matrix t
 #'
 #' @export
 #' @examples
-#' obj <- dummy_data(output="InMemoryAnnData")
+#' h5ad_file <- system.file("extdata", "example.h5ad", package = "anndataR")
+#' ad <- anndataR::read_h5ad(h5ad_file)
 #' to_Seurat(obj) 
 to_Seurat <- function(obj, 
-                      use_layer=NULL,
-                      ...) { # nolint
+                      counts = c("X",obj$layer),
+                      ...) { # nolint 
   requireNamespace("SeuratObject")
   
   stopifnot(inherits(obj, "AbstractAnnData"))
@@ -78,7 +80,7 @@ to_Seurat <- function(obj,
   
   # add layers
   # trackstatus: class=Seurat, feature=get_layers, status=wip
-  #### Store as assay data matrices ####
+  ## Store as assay data matrices
   if(all(seurat_slots %in% obj$layers_keys())){
     message("Setting layers as AssayData matrices.")
     for(key in seurat_slots){
@@ -86,7 +88,7 @@ to_Seurat <- function(obj,
                                                slot = key,
                                                new.data = t(obj$layers[[key]])) 
     }
-  #### Store as multiple assays ####
+  ## Store as multiple assays
   } else {
     message("Setting layers as new assays.")
     for (key in obj$layers_keys()) {
@@ -95,7 +97,7 @@ to_Seurat <- function(obj,
       seurat_obj[[key]] <- SeuratObject::CreateAssayObject(counts = layer_)
     }
   } 
-  #### Add DimReducObjects ####
+  ## Add DimReducObjects
   if(all(c("obsm_keys","varm_keys") %in% names(obj)) &&
      !is.null(obj$obsm_keys()) &&
      !is.null(obj$varm_keys()) ){  
@@ -108,7 +110,7 @@ to_Seurat <- function(obj,
                                                   key=key)
     } 
   }  
-  #### Return ####
+  ## Return 
   seurat_obj
 }
 
@@ -129,22 +131,55 @@ to_Seurat <- function(obj,
 #' @rdname Seurat
 #'
 #' @description `from_Seurat()` converts a Seurat object to an AnnData object.
+#' Only one assay can be converted at a time.
 #'
 #' @param seurat_obj An object inheriting from Seurat.
-#'
-#' @inheritParams dummy_anndata
-#' 
-#' @returns \link[anndataR]{InMemoryAnnData} or \link[anndataR]{HDF5AnnData}
+#' @param output_class Name of the AnnData class. 
+#' Must be one of `"HDF5AnnData"` or `"InMemoryAnnData"`.
+#' @param assay Assay to be converted. If NULL, `DefaultAssay()` is used.
+#' @param X Which of 'counts', 'data', or 'scale.data' will be used for X. 
+#' By default, 'counts' will be used (if it is
+#'  not empty), followed by 'data', then 'scale.data'. 
+#'  The remaining non-empty slots will be stored in different layers.
+#' @param ... Additional arguments passed to the generator function.
 #'
 #' @export
-# TODO: Add parameter to choose which how counts, data and scaled.data are translated into X and layers
 # TODO: add tests with Seurat objects not created by anndataR
 from_Seurat <- function(seurat_obj, 
-                        output_class = c("InMemoryAnnData",
+                        output_class = c("InMemoryAnnData", 
                                          "HDF5AnnData"), 
-                        ...) { # nolint
+                        assay = NULL,
+                        X = "counts", ...) { # nolint
   
   stopifnot(inherits(seurat_obj, "Seurat"))
+  
+  if (!is.null(X)) {
+    if (!X %in% c("counts", "data", "scale.data")) {
+      stop("X must be NULL or one of: 'counts', 'data', 'scale.data'")
+    }
+  }
+  
+  # If a specific assay is selected, use it
+  if (!is.null(assay)) {
+    if (!assay %in% names(seurat_obj@assays)) {
+      stop("'assay' must be NULL or one of: ",
+           paste0("'", names(seurat_obj@assays), "'", collapse = ", "))
+    }
+    assay_name <- assay
+  } else {
+    assay_name <- SeuratObject::DefaultAssay(seurat_obj)
+    
+    # If Seurat object contains multiple assays, 
+    # notify user the Default one is used.
+    if (length(names(seurat_obj@assays)) > 1) {
+      message(
+        "There are ", length(names(seurat_obj@assays)), 
+        " assays in the Seurat object; using the default assay ('",
+        assay_name, 
+        "'). You can use the `assay` parameter to select a specific assay."
+      )
+    }
+  }
   
   # get obs_names
   # trackstatus: class=Seurat, feature=set_obs_names, status=done
@@ -157,11 +192,11 @@ from_Seurat <- function(seurat_obj,
   
   # construct var_names
   # trackstatus: class=Seurat, feature=set_var_names, status=done
-  var_names <- rownames(seurat_obj)
+  var_names <- rownames(seurat_obj@assays[[assay_name]])
   
   # construct var
   # trackstatus: class=Seurat, feature=set_var, status=done
-  var <- seurat_obj@assays[[seurat_obj@active.assay]]@meta.features
+  var <- seurat_obj@assays[[assay_name]]@meta.features
   rownames(var) <- NULL
   
   # use generator to create new AnnData object
@@ -174,33 +209,38 @@ from_Seurat <- function(seurat_obj,
     ...
   )
   
-  # trackstatus: class=Seurat, feature=set_X, status=wip
-  # trackstatus: class=Seurat, feature=set_layers, status=wip
-  for (assay_name in names(seurat_obj@assays)) {
-    # TODO: Maybe we shouldn't use counts but instead data
-    assay_data <- SeuratObject::GetAssayData(seurat_obj, "counts", assay = assay_name)
-    
-    if (nrow(assay_data) != length(var_names) || !identical(rownames(assay_data), var_names)) {
-      warning(
-        "Skipping assay '", assay_name, "' because it has different feature names ",
-        "than the active assay ('", seurat_obj@active.assay, "')."
-      )
-      next
-    }
-    if (ncol(assay_data) != length(obs_names) || !identical(colnames(assay_data), obs_names)) {
-      warning(
-        "Skipping assay '", assay_name, "' because it has different cell names ",
-        "than the active assay ('", seurat_obj@active.assay, "')."
-      )
-      next
+  if (!is.null(X)) {
+    # Check if the slot is not empty
+    if (all(dim(SeuratObject::GetAssayData(seurat_obj, 
+                                           slot = X, 
+                                           assay = assay_name)
+                ) == 0)) {
+      stop("The '", X, "' slot is empty.")
     }
     
-    # remove names
+    assay_data <- SeuratObject::GetAssayData(seurat_obj, 
+                                             slot = X, 
+                                             assay = assay_name)
+    
+    # Remove names
     dimnames(assay_data) <- list(NULL, NULL)
-    if (assay_name == seurat_obj@active.assay) {
-      ad$X <- Matrix::t(assay_data)
-    } else {
-      ad$layers[[assay_name]] <- Matrix::t(assay_data)
+    ad$X <- Matrix::t(assay_data)
+  } else {
+    # Cannot compare other values with NULL
+    X <- "none"
+  }
+  
+  # Add the remaining non-empty slots as layers
+  slots <- c("counts", "data", "scale.data")
+  slots <- slots[slots != X]
+  
+  for (slot in slots) {
+    if (!all(dim(SeuratObject::GetAssayData(seurat_obj, slot = slot)) == 0)) {
+      assay_data <- SeuratObject::GetAssayData(seurat_obj, 
+                                               slot = slot, 
+                                               assay = assay_name)
+      dimnames(assay_data) <- list(NULL, NULL)
+      ad$layers[[slot]] <- Matrix::t(assay_data)
     }
   }