+++ tests/testthat/test-HDF5-read.R
@@ -1,37 +1,37 @@
 skip_if_not_installed("rhdf5")
 
-file <- system.file("extdata", "example.h5ad", package = "anndataR")
+h5ad_file <- system.file("extdata", "example.h5ad", package = "anndataR")
 
 test_that("reading encoding works", {
-  encoding <- read_h5ad_encoding(file, "obs")
+  encoding <- read_h5ad_encoding(h5ad_file, "obs")
   expect_equal(names(encoding), c("type", "version"))
 })
 
 test_that("reading dense matrices works", {
-  mat <- read_h5ad_dense_array(file, "layers/dense_counts")
+  mat <- read_h5ad_dense_array(h5ad_file, "layers/dense_counts")
   expect_true(is.matrix(mat))
   expect_type(mat, "integer")
   expect_equal(dim(mat), c(50, 100))
 
-  mat <- read_h5ad_dense_array(file, "layers/dense_X")
+  mat <- read_h5ad_dense_array(h5ad_file, "layers/dense_X")
   expect_true(is.matrix(mat))
   expect_type(mat, "double")
   expect_equal(dim(mat), c(50, 100))
 })
 
 test_that("reading sparse matrices works", {
-  mat <- read_h5ad_sparse_array(file, "layers/csc_counts", type = "csc")
+  mat <- read_h5ad_sparse_array(h5ad_file, "layers/csc_counts", type = "csc")
   expect_s4_class(mat, "dgCMatrix")
   expect_equal(dim(mat), c(50, 100))
 
-  mat <- read_h5ad_sparse_array(file, "layers/counts", type = "csr")
+  mat <- read_h5ad_sparse_array(h5ad_file, "layers/counts", type = "csr")
   expect_s4_class(mat, "dgRMatrix")
   expect_equal(dim(mat), c(50, 100))
 })
 
 test_that("reading recarrays works", {
   array_list <- read_h5ad_rec_array(
-    file, "uns/rank_genes_groups/logfoldchanges"
+    h5ad_file, "uns/rank_genes_groups/logfoldchanges"
   )
   expect_true(is.list(array_list))
   expect_equal(names(array_list), c("0", "1", "2", "3", "4", "5"))
@@ -43,66 +43,66 @@ test_that("reading recarrays works", {
 })
 
 test_that("reading 1D numeric arrays works", {
-  array_1d <- read_h5ad_dense_array(file, "obs/Int")
+  array_1d <- read_h5ad_dense_array(h5ad_file, "obs/Int")
   expect_vector(array_1d, ptype = integer(), size = 50)
 
-  array_1d <- read_h5ad_dense_array(file, "obs/Float")
+  array_1d <- read_h5ad_dense_array(h5ad_file, "obs/Float")
   expect_vector(array_1d, ptype = double(), size = 50)
 })
 
 test_that("reading 1D sparse numeric arrays works", {
-  array_1d <- read_h5ad_sparse_array(file, "uns/Sparse1D", type = "csc")
+  array_1d <- read_h5ad_sparse_array(h5ad_file, "uns/Sparse1D", type = "csc")
   expect_s4_class(array_1d, "dgCMatrix")
   expect_equal(dim(array_1d), c(1, 6))
 })
 
 test_that("reading 1D nullable arrays works", {
-  array_1d <- read_h5ad_nullable_integer(file, "obs/IntNA")
+  array_1d <- read_h5ad_nullable_integer(h5ad_file, "obs/IntNA")
   expect_vector(array_1d, ptype = integer(), size = 50)
   expect_true(any(is.na(array_1d)))
 
-  array_1d <- read_h5ad_dense_array(file, "obs/FloatNA")
+  array_1d <- read_h5ad_dense_array(h5ad_file, "obs/FloatNA")
   expect_vector(array_1d, ptype = double(), size = 50)
   expect_true(any(is.na(array_1d)))
 
-  array_1d <- read_h5ad_nullable_boolean(file, "obs/Bool")
+  array_1d <- read_h5ad_nullable_boolean(h5ad_file, "obs/Bool")
   expect_vector(array_1d, ptype = logical(), size = 50)
   expect_false(any(is.na(array_1d)))
 
-  array_1d <- read_h5ad_nullable_boolean(file, "obs/BoolNA")
+  array_1d <- read_h5ad_nullable_boolean(h5ad_file, "obs/BoolNA")
   expect_vector(array_1d, ptype = logical(), size = 50)
   expect_true(any(is.na(array_1d)))
 })
 
 test_that("reading string scalars works", {
-  scalar <- read_h5ad_string_scalar(file, "uns/StringScalar")
+  scalar <- read_h5ad_string_scalar(h5ad_file, "uns/StringScalar")
   expect_equal(scalar, "A string")
 })
 
 test_that("reading numeric scalars works", {
-  scalar <- read_h5ad_numeric_scalar(file, "uns/IntScalar")
+  scalar <- read_h5ad_numeric_scalar(h5ad_file, "uns/IntScalar")
   expect_equal(scalar, 1)
 })
 
 test_that("reading string arrays works", {
-  array <- read_h5ad_string_array(file, "uns/String")
+  array <- read_h5ad_string_array(h5ad_file, "uns/String")
   expect_vector(array, ptype = character(), size = 10)
   expect_equal(array[3], "String 2")
 
-  array <- read_h5ad_string_array(file, "uns/String2D")
+  array <- read_h5ad_string_array(h5ad_file, "uns/String2D")
   expect_true(is.matrix(array))
   expect_type(array, "character")
   expect_equal(dim(array), c(5, 10))
 })
 
 test_that("reading mappings works", {
-  mapping <- read_h5ad_mapping(file, "uns")
+  mapping <- read_h5ad_mapping(h5ad_file, "uns")
   expect_type(mapping, "list")
   expect_type(names(mapping), "character")
 })
 
 test_that("reading dataframes works", {
-  df <- read_h5ad_data_frame(file, "obs")
+  df <- read_h5ad_data_frame(h5ad_file, "obs")
   expect_s3_class(df, "data.frame")
   expect_equal(
     colnames(df),
@@ -117,13 +117,14 @@ test_that("reading dataframes works", {
 test_that("reading H5AD as SingleCellExperiment works", {
   skip_if_not_installed("SingleCellExperiment")
 
-  sce <- read_h5ad(file, to = "SingleCellExperiment")
+  sce <- read_h5ad(h5ad_file, to = "SingleCellExperiment")
   expect_s4_class(sce, "SingleCellExperiment")
 })
 
 test_that("reading H5AD as Seurat works", {
   skip_if_not_installed("SeuratObject")
 
-  seurat <- read_h5ad(file, to = "Seurat")
+  # TODO: remove this suppression when the to_seurat, from_seurat functions are updated.
+  seurat <- suppressWarnings(read_h5ad(h5ad_file, to = "Seurat"))
   expect_s4_class(seurat, "Seurat")
 })